+++ ohsome_quality_api/geodatabase/select_building_area.sql
@@ -1,6 +1,6 @@
 WITH bpoly AS (
     SELECT
-        ST_Setsrid (ST_GeomFromGeoJSON ($1), 4326) AS geom
+        ST_Setsrid (ST_GeomFromGeoJSON (%s), 4326) AS geom
 )
 SELECT
     SUM(eubucco.area) as area
@@ -8,4 +8,4 @@ FROM
     eubucco,
     bpoly
 WHERE
-    ST_Intersects (eubucco.geom, bpoly.geom);
+    ST_Intersects (eubucco.centroid, bpoly.geom);