+++ ohsome_quality_api/indicators/building_comparison/indicator.py
@@ -1,12 +1,15 @@
 import logging
+import os
 from string import Template
 
 import geojson
 import plotly.graph_objects as pgo
+import psycopg
 from dateutil import parser
 from geojson import Feature, MultiPolygon, Polygon
 from numpy import mean
 
+from ohsome_quality_api.config import get_config_value
 from ohsome_quality_api.definitions import Color, get_attribution
 from ohsome_quality_api.geodatabase import client as db_client
 from ohsome_quality_api.indicators.base import BaseIndicator
@@ -54,28 +57,28 @@ class BuildingComparison(BaseIndicator):
         else:
             self.coverage["EUBUCCO"] = None
             return
-        if not self.check_major_edge_cases():
-            self.result.description = ""
-            self.feature = await db_client.get_eubucco_coverage_intersection(
-                self.feature
-            )
-            db_query_result = await db_client.get_building_area(self.feature)
-            raw = db_query_result[0]["area"] or 0
-            self.area_references["EUBUCCO"] = raw / (1000 * 1000)
 
-            osm_query_result = await ohsome_client.query(
-                self.topic,
-                self.feature,
-            )
-            raw = osm_query_result["result"][0]["value"] or 0  # if None
-            self.area_osm = raw / (1000 * 1000)
-            self.result.timestamp_osm = parser.isoparse(
-                osm_query_result["result"][0]["timestamp"]
-            )
+        edge_case = self.check_major_edge_cases()
+        if edge_case:
+            self.result.description = edge_case
+            return
+
+        self.feature = await db_client.get_eubucco_coverage_intersection(self.feature)
+        db_query_result = await get_eubucco_building_area(self.feature)
+        self.area_references["EUBUCCO"] = db_query_result / (1000 * 1000)
+        osm_query_result = await ohsome_client.query(
+            self.topic,
+            self.feature,
+        )
+        raw = osm_query_result["result"][0]["value"] or 0  # if None
+        self.area_osm = raw / (1000 * 1000)
+        self.result.timestamp_osm = parser.isoparse(
+            osm_query_result["result"][0]["timestamp"]
+        )
 
     def calculate(self) -> None:
         # TODO: put checks into check_corner_cases. Let result be undefined.
-        if not self.result.description == "":
+        if self.result.label == "undefined" and self.check_major_edge_cases():
             return
         if self.check_minor_edge_cases():
             self.result.description = self.check_minor_edge_cases()
@@ -101,9 +104,11 @@ class BuildingComparison(BaseIndicator):
         elif self.th_low > self.result.value >= 0:
             self.result.class_ = 1
         elif self.result.value > self.above_one_th:
+            # TODO: move this to edge_case functions
             self.result.description += (
-                "Warning: No quality estimation made. "
-                "OSM and reference data differ. Reference data is likely outdated. "
+                "Warning: Because of a big difference between OSM and the reference "
+                + "data no quality estimation has been made. "
+                + "It could be that the reference data is outdated. "
             )
 
         template = Template(self.metadata.result_description)
@@ -116,7 +121,10 @@ class BuildingComparison(BaseIndicator):
 
     def create_figure(self) -> None:
         if self.result.label == "undefined" and self.check_major_edge_cases():
-            logging.info("Result is undefined. Skipping figure creation.")
+            logging.info(
+                "Result is undefined and major edge case is present. "
+                + "Skipping figure creation."
+            )
             return
         fig = pgo.Figure()
         fig.add_trace(
@@ -140,33 +148,32 @@ class BuildingComparison(BaseIndicator):
         fig.update_layout(title_text=("Building Comparison"), showlegend=True)
         fig.update_yaxes(title_text="Building Area [km²]")
         fig.update_xaxes(
-            title_text="Datasets (" + get_sources(self.area_references.keys()) + ")"
+            title_text="Reference Datasets ("
+            + get_sources(self.area_references.keys())
+            + ")"
         )
         raw = fig.to_dict()
         raw["layout"].pop("template")  # remove boilerplate
         self.result.figure = raw
 
-    def check_major_edge_cases(self) -> bool:
+    def check_major_edge_cases(self) -> str:
+        """If edge case is present return description if not return empty string."""
         coverage = self.coverage["EUBUCCO"]
-        # TODO: generalize function
         if coverage is None or coverage == 0.00:
-            self.result.description = (
-                "Reference dataset does not cover area-of-interest."
-            )
-            return True
+            return "Reference dataset does not cover area-of-interest."
         elif coverage < 0.10:
-            self.result.description = (
+            return (
                 "Only {:.2f}% of the area-of-interest is covered ".format(
                     coverage * 100
                 )
                 + "by the reference dataset (EUBUCCO). "
                 + "No quality estimation is possible."
             )
-            return True
         else:
-            return False
+            return ""
 
     def check_minor_edge_cases(self) -> str:
+        """If edge case is present return description if not return empty string."""
         coverage = self.coverage["EUBUCCO"]
         if coverage < 0.95:
             return (
@@ -176,6 +183,28 @@ class BuildingComparison(BaseIndicator):
             )
         else:
             return ""
+
+
+async def get_eubucco_building_area(bpoly: Feature) -> float:
+    """Get the building area for a AoI from the EUBUCCO dataset."""
+    # TODO: https://github.com/GIScience/ohsome-quality-api/issues/746
+    file_path = os.path.join(db_client.WORKING_DIR, "select_building_area.sql")
+    with open(file_path, "r") as file:
+        query = file.read()
+    geom = str(bpoly.geometry)
+    dns = "postgres://{user}:{password}@{host}:{port}/{database}".format(
+        host=get_config_value("postgres_host"),
+        port=get_config_value("postgres_port"),
+        database=get_config_value("postgres_db"),
+        user=get_config_value("postgres_user"),
+        password=get_config_value("postgres_password"),
+    )
+    with psycopg.connect(dns) as con:
+        with con.cursor() as cur:
+            cur.execute(query, (geom,))
+            res = cur.fetchall()
+            con.commit()
+    return res[0][0] or 0.0
 
 
 def get_sources(reference_datasets):