+++ tests/integrationtests/indicators/test_building_comparison.py
@@ -15,27 +15,27 @@ from tests.integrationtests.utils import oqapi_vcr
 
 @pytest.fixture(scope="class")
 def mock_get_building_area(class_mocker):
-    async_mock = AsyncMock(return_value=[{"area": 4842587.791645115}])
+    async_mock = AsyncMock(return_value=4842587.791645115)
     class_mocker.patch(
-        "ohsome_quality_api.indicators.building_completeness.indicator.db_client.get_building_area",
+        "ohsome_quality_api.indicators.building_comparison.indicator.get_eubucco_building_area",
         side_effect=async_mock,
     )
 
 
 @pytest.fixture(scope="class")
 def mock_get_building_area_low(class_mocker):
-    async_mock = AsyncMock(return_value=[{"area": 1}])
+    async_mock = AsyncMock(return_value=1)
     class_mocker.patch(
-        "ohsome_quality_api.indicators.building_completeness.indicator.db_client.get_building_area",
+        "ohsome_quality_api.indicators.building_comparison.indicator.get_eubucco_building_area",
         side_effect=async_mock,
     )
 
 
 @pytest.fixture(scope="class")
 def mock_get_building_area_empty(class_mocker):
-    async_mock = AsyncMock(return_value=[{"area": 0}])
+    async_mock = AsyncMock(return_value=0)
     class_mocker.patch(
-        "ohsome_quality_api.indicators.building_completeness.indicator.db_client.get_building_area",
+        "ohsome_quality_api.indicators.building_comparison.indicator.get_eubucco_building_area",
         side_effect=async_mock,
     )
 
@@ -44,7 +44,7 @@ def mock_get_building_area_empty(class_mocker):
 def mock_get_eubucco_coverage_intersection(class_mocker, feature_germany_berlin):
     async_mock = AsyncMock(return_value=feature_germany_berlin)
     class_mocker.patch(
-        "ohsome_quality_api.indicators.building_completeness.indicator.db_client.get_eubucco_coverage_intersection",
+        "ohsome_quality_api.indicators.building_comparison.indicator.db_client.get_eubucco_coverage_intersection",
         side_effect=async_mock,
     )
 
@@ -53,7 +53,7 @@ def mock_get_eubucco_coverage_intersection(class_mocker, feature_germany_berlin)
 def mock_get_eubucco_coverage_intersection_area(class_mocker):
     async_mock = AsyncMock(return_value=[{"area_ratio": 1}])
     class_mocker.patch(
-        "ohsome_quality_api.indicators.building_completeness.indicator.db_client.get_eubucco_coverage_intersection_area",
+        "ohsome_quality_api.indicators.building_comparison.indicator.db_client.get_eubucco_coverage_intersection_area",
         side_effect=async_mock,
     )
 
@@ -138,8 +138,9 @@ class TestCalculate:
         assert indicator.result.class_ is None
         assert indicator.result.description is not None
         assert (
-            "Warning: No quality estimation made. "
-            "OSM and reference data differ. Reference data is likely outdated."
+            "Warning: Because of a big difference between OSM and the reference "
+            + "data no quality estimation has been made. "
+            + "It could be that the reference data is outdated. "
             in indicator.result.description
         )
         assert indicator.result.label == "undefined"